Backend/Project-Service/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
python-decouple
pytz
sqlparse
mysqlclient
django-mysql
pika
